feat: move and rename recovery; drop reference to protection.
logion-network/logion-internal#1197
