+++ src/legal-officer/recovery/RecoveryDetails.test.tsx
@@ -1,9 +1,9 @@
-jest.mock('../logion-chain/Signature');
-jest.mock("../common/CommonContext");
-jest.mock("../loc/Model");
-jest.mock("./Model");
-jest.mock("./LegalOfficerContext");
-jest.mock("../logion-chain");
+jest.mock('../../logion-chain/Signature');
+jest.mock("../../common/CommonContext");
+jest.mock("../../loc/Model");
+jest.mock("../Model");
+jest.mock("../LegalOfficerContext");
+jest.mock("../../logion-chain");
 
 import { LegalOfficerCase } from "@logion/node-api";
 import { SubmittableExtrinsic } from "@polkadot/api-base/types";
@@ -11,15 +11,15 @@ import { render, waitFor, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 
 import RecoveryDetails from './RecoveryDetails';
-import { RecoveryInfo } from './Types';
-import { acceptProtectionRequest, rejectProtectionRequest } from '../loc/__mocks__/ModelMock';
-import { setFetchRecoveryInfo } from './__mocks__/ModelMock';
-import { PROTECTION_REQUESTS_HISTORY } from './TestData';
-import { axiosMock, setAddresses, DEFAULT_LEGAL_OFFICER_ACCOUNT } from '../logion-chain/__mocks__/LogionChainMock';
-import { setParams, navigate } from '../__mocks__/ReactRouterMock';
-import { refreshRequests } from './__mocks__/LegalOfficerContextMock';
+import { RecoveryInfo } from '../Types';
+import { acceptProtectionRequest, rejectProtectionRequest } from '../../loc/__mocks__/ModelMock';
+import { setFetchRecoveryInfo } from '../__mocks__/ModelMock';
+import { PROTECTION_REQUESTS_HISTORY } from '../TestData';
+import { axiosMock, setAddresses, DEFAULT_LEGAL_OFFICER_ACCOUNT } from '../../logion-chain/__mocks__/LogionChainMock';
+import { setParams, navigate } from '../../__mocks__/ReactRouterMock';
+import { refreshRequests } from '../__mocks__/LegalOfficerContextMock';
 import { It, Mock } from 'moq.ts';
-import { mockValidPolkadotAccountId, setupApiMock, CLOSED_IDENTITY_LOC_ID } from 'src/__mocks__/LogionMock';
+import { mockValidPolkadotAccountId, setupApiMock, CLOSED_IDENTITY_LOC_ID } from '../../__mocks__/LogionMock';
 
 describe("RecoveryDetails", () => {
