+++ src/legal-officer/recovery/RecoveryRequestStatus.css
@@ -1,4 +1,4 @@
-.ProtectionRequestStatus {
+.RecoveryRequestStatus {
     display: inline-block;
     padding: 10px;
     border: 1px solid white;
