+++ src/legal-officer/recovery/RecoveryRequests.test.tsx
@@ -1,8 +1,8 @@
-jest.mock("../common/CommonContext");
+jest.mock("../../common/CommonContext");
 
 import React from 'react';
 import RecoveryRequests from './RecoveryRequests';
-import { shallowRender } from '../tests';
+import { shallowRender } from '../../tests';
 
 test('renders', () => {
     const tree = shallowRender(<RecoveryRequests />);
