+++ src/legal-officer/recovery/RecoveryRequestsHistory.test.tsx
@@ -1,10 +1,10 @@
-jest.mock('./LegalOfficerContext');
-jest.mock('./Model');
+jest.mock('../LegalOfficerContext');
+jest.mock('../Model');
 
-import { shallowRender } from '../tests';
+import { shallowRender } from '../../tests';
 import RecoveryRequestsHistory from './RecoveryRequestsHistory';
-import { setRecoveryRequestsHistory } from './__mocks__/LegalOfficerContextMock';
-import { RECOVERY_REQUESTS_HISTORY } from './TestData';
+import { setRecoveryRequestsHistory } from '../__mocks__/LegalOfficerContextMock';
+import { RECOVERY_REQUESTS_HISTORY } from '../TestData';
 
 test("Renders null with no data", () => {
     const tree = shallowRender(<RecoveryRequestsHistory />);
